control hardware test refactors
