+++ control-board/src/bin/hwtest-motor/pins.rs
